requirements.txt
numpy==1.21.4
opencv_python==4.5.4.60
pandas==1.3.4
pymilvus==1.1.2
PyMySQL==1.0.2
streamlit==1.2.0
torch==1.10.0
torchvision==0.11.1
